fix scale gizmo vao caching
