+++ izzet/client/clientapi.go
@@ -699,7 +699,7 @@ func (g *Client) BuildNavMesh(app renderiface.App, iterationCount int, walkableH
 				tv2 := v3.Sub(v2)
 
 				normal := tv1.Cross(tv2).Normalize()
-				isUp := normal.Dot(up) >= 0.9
+				isUp := normal.Dot(up) >= 0.7
 
 				navmesh.RasterizeTriangle2(v1, v2, v3, 1, 1, 1, hf, isUp, climbableHeight)
 			}
