+++ izzet/render/shaders.go
@@ -1,30 +1,8 @@
 package render
 
 import (
-	"github.com/go-gl/gl/v3.2-core/gl"
 	"github.com/kkevinchou/kitolib/shaders"
 )
-
-type Platform interface {
-	NewFrame()
-	DisplaySize() [2]float32
-	FramebufferSize() [2]float32
-}
-
-func initOpenGLRenderSettings() {
-	// gl.ClearColor(0.0, 0.5, 0.5, 0.0)
-	gl.ClearColor(1, 1, 1, 1)
-	gl.ClearDepth(1)
-	gl.Enable(gl.DEPTH_TEST)
-	gl.DepthFunc(gl.LEQUAL)
-	gl.Enable(gl.CULL_FACE)
-	gl.CullFace(gl.BACK)
-	gl.FrontFace(gl.CCW)
-	gl.Enable(gl.MULTISAMPLE)
-	gl.Enable(gl.BLEND)
-	gl.BlendFunc(gl.SRC_ALPHA, gl.ONE_MINUS_SRC_ALPHA)
-	gl.Disable(gl.FRAMEBUFFER_SRGB)
-}
 
 func compileShaders(shaderManager *shaders.ShaderManager) {
 	if err := shaderManager.CompileShaderProgram("skybox", "skybox", "skybox", ""); err != nil {
